refactor(id)!: Update id type, session id name and use consistently
BREAKING CHANGE: Make vaccel_id_t a fixed-size type (int64), rename id
variable in vaccel_session to match the resource convention/type and
update types where needed:
- Change vaccel_id_t from `long long int` to `int64_t` for portability
- Rename `session_id` to `id` in `struct vaccel_session` following the
  resource naming convention
- Make id types `vaccel_id_t` in `struct vaccel_session`. This is the
  type that `id_pool` returns and what is used in vaccel_resource
- Update variables and print format strings in all related files
- Move id generation at the start of `vaccel_sess_init()` so it will
  either fail early if we can't generate one or it will be available for
  the virtio plugins

PR: https://github.com/nubificus/vaccel/pull/137
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>

+++ src/include/id.h
@@ -4,4 +4,4 @@
 
 #include <stdint.h>
 
-typedef long long int vaccel_id_t;
+typedef int64_t vaccel_id_t;
